# docs/swagger.yaml
openapi: 3.0.2
info:
    title: games-api
    version: 1.0.0
    description: ''
paths:
    /games:
        get:
            tags:
                - games
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items: {}
                            examples:
                                List of games:
                                    value:
                                        - Witcher 3
                                        - Cyberpunk 2077
                                        - Red Dead Redemption 2
                                        - Elden Ring
                    description: Returns full details of a given id game
            summary: Get a list of all games
            description: Returns a list of all games in API
    '/games/{id}':
        get:
            tags:
                - details
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items: {}
                            examples:
                                Full details of a given game:
                                    value: |4-
                                            {id: 1, name: "Witcher 3", price: 29.99},
                                            {id: 2, name: "Cyberpunk 2077", price: 59.99},
                                            {id: 3, name: "Red Dead Redemption 2", price: 24.99},
                                            {id: 4, name: "Elden Ring", price: 20.99},
                                            {id: 5, name: "Roblox", price: 0},
                                            {id: 6, name: "Outlast", price: 69.99},
                                            {id: 7, name: "Valorant", price: 0},
                                            {id: 8, name: "Silent Hill", price: 49.99}
                    description: Returns full details of a given id game
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Game /w+ was not found:
                                    value:
                                        error: Game \w+ was not found in game array.
                    description: Game was not found in games array
            operationId: id
            summary: Get full details of a game
            description: |
                Returns full details of a given id game
components:
    schemas:
        error:
            title: Root Type for error
            description: ''
            type: object
            properties:
                error:
                    type: string
            example:
                error: Error message.